+++ app/Http/Controllers/Admin/PostController.php
@@ -8,17 +8,59 @@ use App\Models\Post;
 use App\Http\Resources\PostResource;
 use App\Http\Requests\StorePostRequest;
 use App\Http\Requests\UpdatePostRequest;
+use App\Models\Category;
+use Illuminate\Http\Request;
 
 class PostController extends Controller
 {
+     /**
+     * Display post by category of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+
+     public function getPostByCategpory(Request $request)
+     {
+        $category_id = $request->query('category_id');
+
+        $post = Post::whereHas('category', function($query) use($category_id) {
+            $query->where('category_id', $category_id);
+        })->get();
+
+        if(!$post) {
+            return response()->json('Post category Not found');
+        }
+
+        return response()->json($post);
+     }
     /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
-        $posts = Post::all();
+        $post_query = Post::with('category');
+
+       
+
+        if($request->keyword) {
+            $post_query->where('name', 'LIKE', '%', $request->keyword, '%');
+        }
+
+        if($request->category) {
+            $post_query->whereHas('category', function($query) use($request) {
+                $query->where('name', $request->category);
+            });
+        }
+
+         $posts = $post_query->orderBy('id', 'desc')->get();
+
+        return response()->json([
+            'message' => 'blog fetch successfully',
+            'post' => $posts
+        ]);
+
 
         return PostResource::collection($posts);
     }
@@ -73,8 +115,14 @@ class PostController extends Controller
      * @param  \App\Models\Post  $post
      * @return \Illuminate\Http\Response
      */
-    public function show(Post $post)
+    public function show($id)
     {
+        $post = Post::find($id);
+
+        if(!$post) {
+            return response()->json('post not found');
+        }
+
         return new PostResource($post);
     }
 
@@ -96,8 +144,14 @@ class PostController extends Controller
      * @param  \App\Models\Post  $post
      * @return \Illuminate\Http\Response
      */
-    public function update(UpdatePostRequest $request, Post $post)
+    public function update(UpdatePostRequest $request,  $id)
     {
+        $post = Post::find($id);
+
+        if(!$post) {
+            return response()->json('post not found');
+        }
+
         if( $request->hasFile('image')) {
   
             $image = $request->image;
@@ -111,11 +165,11 @@ class PostController extends Controller
             $post->image = 'posts/image/' . $image_new_name;
       }
 
-        $post->category_id = $request->input('category_id');
-        $post->title = $request->input('title');
-        $post->description = $request->input('description');
-        $post->views = $request->input('views');
-        $post->date = $request->input('date');
+        $post->category_id = $request->category_id;
+        $post->title = $request->title;
+        $post->description = $request->description;
+        $post->views = $request->views;
+        $post->date = $request->date;
         $post->update();
 
         return new PostResource($post);
@@ -127,8 +181,14 @@ class PostController extends Controller
      * @param  \App\Models\Post  $post
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Post $post)
+    public function destroy($id)
     {
+        $post = Post::find($id);
+
+        if(!$post) {
+            return response()->json('post not found');
+        }
+
         $post = $post->delete();
 
         return response()->json([
